Crate polkavm_derive

The derive crate for PolkaVM guest programs

Modules

default_abi

Macros

min_stack_size

Sets the minimum stack size.

Structs

LeakingAllocator

A basic memory allocator which doesn’t support deallocation.

Functions

heap_base

memset⚠

A hardware accelerated memset.

sbrk

Increases the size of the program’s heap by a given number of bytes, allocating memory if necessary. If successful returns a pointer to the end of the heap. If unsuccessful returns a null pointer.

Attribute Macros

polkavm_define_abi

polkavm_export

polkavm_import